pyproject.toml
[project]
name = "notely"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "markdown-it-py>=4.0.0",
    "markdownify>=1.2.2",
    "nicegui>=3.9.0",
    "pyyaml>=6.0.3",
    "rich-click>=1.9.7",
]

[project.scripts]
notely = "notely.notely:main"

[dependency-groups]
dev = [
    "hypothesis>=6.151.9",
    "pytest>=9.0.2",
    "pytest-sugar>=1.1.1",
    "pytest-custom-exit-code>=0.3.0",
    "pytest-order>=1.3.0",
    "pytest-parametrized>=1.7",
    "pytest-randomly>=3.16.0",
    "pytest-repeat>=0.9.4",
    "pytest-xdist>=3.6.1",
    "python-lsp-server>=1.14.0",
]
lint = [
    "ruff>=0.15.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [ "notely" ]

[tool.pytest.ini_options]
xfail_strict = true
addopts = "--dist loadgroup -n auto --order-group-scope class -rs --strict-markers --suppress-no-test-exit-code -vvv -x"
